Upgrades to v4 guides and docs. WIP
