+++ posts/docs/v4/api/link.md
@@ -1,0 +1,7 @@
+---
+title: nextein/link
+order: 30
+deprecated: true
+---
+
+This has been deprecated in version 4. Use next.js link instead.
