+++ posts/docs/v4/api/post.md
@@ -1,0 +1,7 @@
+---
+title: nextein/post
+order: 20
+deprecated: true
+---
+
+This has been deprecated in v4. See [fetcher](../fetcher) for `getSaticProps` and `getStaticPaths` usage.
