+++ posts/docs/v4/api/posts.md
@@ -1,0 +1,7 @@
+---
+title: nextein/posts
+order: 10
+deprecated: true
+---
+
+This has been deprecated in v4. See [fetcher](../fetcher) for `getSaticProps` and `getStaticPaths` usage.
