+++ src/site.json
@@ -5,6 +5,7 @@
   
   "versions": {
     "latest": "v3",
+    "v4": "v4.0.0-beta",
     "v3": "v3.2.4",
     "v2": "v2.7.3"
   }
